+++ packages/api/src/client/types/chat/bsky/convo/updateRead.ts
@@ -40,6 +40,16 @@ export interface Response {
   data: OutputSchema
 }
 
+export class InvalidConvoError extends XRPCError {
+  constructor(src: XRPCError) {
+    super(src.status, src.error, src.message, src.headers, { cause: src })
+  }
+}
+
 export function toKnownErr(e: any) {
+  if (e instanceof XRPCError) {
+    if (e.error === 'InvalidConvo') return new InvalidConvoError(e)
+  }
+
   return e
 }