+++ packages/api/src/client/types/chat/bsky/group/editGroup.ts
@@ -1,0 +1,69 @@
+/**
+ * GENERATED CODE - DO NOT MODIFY
+ */
+import { HeadersMap, XRPCError } from '@atproto/xrpc'
+import { type ValidationResult, BlobRef } from '@atproto/lexicon'
+import { CID } from 'multiformats/cid'
+import { validate as _validate } from '../../../../lexicons'
+import {
+  type $Typed,
+  is$typed as _is$typed,
+  type OmitKey,
+} from '../../../../util'
+import type * as ChatBskyConvoDefs from '../convo/defs.js'
+
+const is$typed = _is$typed,
+  validate = _validate
+const id = 'chat.bsky.group.editGroup'
+
+export type QueryParams = {}
+
+export interface InputSchema {
+  convoId: string
+  name: string
+}
+
+export interface OutputSchema {
+  convo: ChatBskyConvoDefs.ConvoView
+}
+
+export interface CallOptions {
+  signal?: AbortSignal
+  headers?: HeadersMap
+  qp?: QueryParams
+  encoding?: 'application/json'
+}
+
+export interface Response {
+  success: boolean
+  headers: HeadersMap
+  data: OutputSchema
+}
+
+export class ConvoLockedError extends XRPCError {
+  constructor(src: XRPCError) {
+    super(src.status, src.error, src.message, src.headers, { cause: src })
+  }
+}
+
+export class InvalidConvoError extends XRPCError {
+  constructor(src: XRPCError) {
+    super(src.status, src.error, src.message, src.headers, { cause: src })
+  }
+}
+
+export class InsufficientRoleError extends XRPCError {
+  constructor(src: XRPCError) {
+    super(src.status, src.error, src.message, src.headers, { cause: src })
+  }
+}
+
+export function toKnownErr(e: any) {
+  if (e instanceof XRPCError) {
+    if (e.error === 'ConvoLocked') return new ConvoLockedError(e)
+    if (e.error === 'InvalidConvo') return new InvalidConvoError(e)
+    if (e.error === 'InsufficientRole') return new InsufficientRoleError(e)
+  }
+
+  return e
+}