+++ packages/ozone/src/lexicon/types/chat/bsky/group/editJoinLink.ts
@@ -1,0 +1,47 @@
+/**
+ * GENERATED CODE - DO NOT MODIFY
+ */
+import { type ValidationResult, BlobRef } from '@atproto/lexicon'
+import { CID } from 'multiformats/cid'
+import { validate as _validate } from '../../../../lexicons'
+import {
+  type $Typed,
+  is$typed as _is$typed,
+  type OmitKey,
+} from '../../../../util'
+import type * as ChatBskyGroupDefs from './defs.js'
+
+const is$typed = _is$typed,
+  validate = _validate
+const id = 'chat.bsky.group.editJoinLink'
+
+export type QueryParams = {}
+
+export interface InputSchema {
+  convoId: string
+  requireApproval?: boolean
+  joinRule?: ChatBskyGroupDefs.JoinRule
+}
+
+export interface OutputSchema {
+  joinLink: ChatBskyGroupDefs.JoinLinkView
+}
+
+export interface HandlerInput {
+  encoding: 'application/json'
+  body: InputSchema
+}
+
+export interface HandlerSuccess {
+  encoding: 'application/json'
+  body: OutputSchema
+  headers?: { [key: string]: string }
+}
+
+export interface HandlerError {
+  status: number
+  message?: string
+  error?: 'InvalidConvo' | 'InsufficientRole' | 'NoJoinLink'
+}
+
+export type HandlerOutput = HandlerError | HandlerSuccess